NeuroMateV1/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.38.0
google-adk[database]==0.3.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
deprecated==1.2.14
flask==3.0.2
gunicorn==21.2.0
vertexai>=0.1.0